Move test-only code to new file, and exclude that from coverage
